fix: Fabrizio review — four audible bugs, honest ABI, real gates
Queue-jumpers (all ear-audible):
- arranger: multi-bar sections never played past bar 1 — the section
  clock restarted at EVERY bar boundary; it now restarts only on wrap
  or actual switch (and the duplicated nested boundary if is gone).
  Latent until the first real style (builtins are one-bar); pinned by
  a 2-bar test style through the new Arranger::load_style(const
  Style*), which compiled user styles will need anyway.
- timeline: gate 0 through the binary ABI put the NoteOff on the
  NoteOn's own tick, where the D29 off-before-on order guarantees a
  stuck note. set_step now refuses vel>0/gate==0 at the core — the
  ABI is the product boundary (D26), the shell check was decoration.
- parser: EOX (F7) neither terminated the message being assembled nor
  cleared running status; 90 3C F7 40 built a phantom note and data
  after F7 re-armed as running status. Both per MIDI 1.0 now, tested.
- timeline/arranger retrigger: re-firing a note whose previous
  NoteOff was still pending violated §9.B (duplicate on) and the
  stale off then truncated the new note. The scheduler grows
  cancel_note_off (earliest matching pending off is tombstoned and
  skipped at pop); pattern scheduling closes the sounding note on the
  re-fire tick. Pinned: ons at 0/240, offs at 240/340, nothing at 300.

Architecture and honesty:
- the arm gate now certifies the product, not 10% of it: new
  engine_checks.cpp instantiates the full Engine and drives the ABI;
  the firmware stub calls it (stub text: 991 B -> ~45 KB of real
  freestanding object code).
- kChordHold answered the ABI and did nothing; it now warns the new
  kUnsupported until the live-keyboard milestone implements it, and
  the dead hold state is gone.
- chord out while sounding stranded the old port: the voicing is
  released on the old destination before switching. ChordSequencer::
  play now refuses while recording (symmetric with start_record).
- shell name tables can no longer diverge from core pools: bounds are
  checked before registering a name the core would refuse.

Structure, tooling, cleanups:
- push_command's 300-line god-switch left the header: engine.cpp with
  per-domain handlers (cmd_transport/routing/chord/seq/track/style),
  ten-line dispatch.
- .clang-tidy gains HeaderFilterRegex 'app/' — the lint gate had
  never judged the ~100%-header core (proof: result.hpp's val_/err_
  passed green; renamed m_val/m_err). Core warnings are now PUBLIC so
  headers are judged by every TU; two sign-conversions clang caught
  where GCC shrugged are fixed.
- jsonl warn table covers all WarnCodes with a static_assert on the
  count (four spelled 'unknown' before); launcher kills get || true;
  alsa_midi double-brace noise removed; vestigial ChordSequencer::
  m_next_step and dead theory arithmetic deleted.

Gates: 23/23 suites (goldens byte-identical), arm cross-build green
with the whole engine linked, lint clean with headers under
judgement, coverage 91.6/98.2/80.7 — all three over the 80% bar.

+++ app/core/include/arrangrr/routing/router.hpp
@@ -77,7 +77,8 @@ class Router {
       }
       MidiMessage out = msg;
       if (has_channel && r.out_channel >= 0) {
-        out.status = static_cast<std::uint8_t>((msg.status & 0xF0u) | (r.out_channel & 0x0Fu));
+        out.status = static_cast<std::uint8_t>((msg.status & 0xF0u) |
+                                               (static_cast<std::uint8_t>(r.out_channel) & 0x0Fu));
       }
       sink(r.out_port, out);
     }
